Thread sort and write values to single file
This is a large commit that covers a few different goals.

First, the project has been successfully altered to only write to a single masterCSV file.

The program will also fully implement a thread safe stack for storing completed files. HOWEVER it has not been implemented in a multithreaded environement yet.

The program also needs work on taking in a column to sort by.

Also the output csv code should be moved to an earlier point in the processing and the File pointer should be assigned to some global file pointer.

As of now the program really works in two stages, 1. sorting the files and 2. merging the files. These should be combined into an overall step using threading in order to increase speed.

The next step should be cleanup code and ensuring the current structure works with all input cases.

Then what follows shoudl be speed optimizations and additional thread creations.